+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Feature
 
-- Add Finney Modal []()
+- Add Finney Modal [](<>)
 
 ### Release v5.0.6
 
@@ -13,7 +13,6 @@
 - Fix block 0 on mobile [#767](https://github.com/MyEtherWallet/MyEtherWallet/pull/767)
 - Update infura endpoint [#763](https://github.com/MyEtherWallet/MyEtherWallet/pull/763)
 - Chrome 72 compatibility for Secalot [#766](https://github.com/MyEtherWallet/MyEtherWallet/pull/766)
-
 
 ### Hotfix v5.0.5-hotfix.1