+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Feature
 
-- Add Finney Modal [](<>)
+- Add Finney Modal [#782](https://github.com/MyEtherWallet/MyEtherWallet/pull/782)
 
 ### Release v5.0.6